#1 called Simplify binary according to the OS type
