+++ bin/simplify.osx
